🐛 fix(document): update API handling for new database structure
With the introduction of the `original` column, we are able to:
- Remove the `type` validation.
- Simplify the query by eliminating the `type` condition from the queries.
