+++ server/api/documents/[id]/index.put.ts
@@ -3,33 +3,24 @@ import { workDocumentSchema } from '~~/types/schema/document';
 
 export default defineEventHandler(async (event) => {
   const id = decodeURI(getRouterParam(event, 'id') || '');
-  const type = decodeURI(getRouterParam(event, 'type') || '').toLowerCase();
   const workDocument = await useValidatedBody(event, workDocumentSchema);
 
-  if (!isValidDocumentType(type)) throw createError({ statusCode: 404 });
   const user = await verifyUserAuthorizationByName(event);
 
   const original = await getWorkDocumentByNameAndId({ name: user.name!, id });
 
   await useDB()
     .insert(tables.documentMitra)
-    .values([
-      {
-        id,
-        type: 'original',
-        value: original,
-        createdAt: new Date(),
-      },
-      {
-        id,
-        type,
-        value: workDocument,
-        createdAt: new Date(),
-      },
-    ])
+    .values({
+      id,
+      original,
+      value: workDocument,
+      createdAt: new Date(),
+    })
     .onConflictDoUpdate({
-      target: [tables.documentMitra.id, tables.documentMitra.type],
+      target: tables.documentMitra.id,
       set: {
+        original,
         value: workDocument,
         isValidated: false,
         isApproved: false,
